Check if a user has access to a list and actoins concerning that list

+++ todolist/views.py
@@ -9,8 +9,6 @@ from .forms import ListForm, TaskForm, TaskImportForm
 from url_shortener.models import UrlDict
 
 
-# middleware
-
 def index(request):
     if request.method == 'GET':
         user: CustomUser = request.user
@@ -25,20 +23,26 @@ def index(request):
 
 
 def list_detail(request, list_id):
-    # check the owner
     if request.method == 'GET':
-        current_list = get_object_or_404(List, id=list_id)  # change the name
-        return render(request, 'todolist/list_detail.html', {'list': current_list})
+        current_list = get_object_or_404(List, id=list_id)
+        user: CustomUser = request.user
+        if user.lists.contains(current_list):
+            return render(request, 'todolist/list_detail.html', {'list': current_list})
+        else:
+            return render(request, 'todolist/access_denied.html')
 
     return render(request, 'todolist/wrong_method.html')
 
 
 def list_delete(request, list_id):
-    # Check if method is DELETE
+    user: CustomUser = request.user
     del_list = get_object_or_404(List, id=list_id)
-    name = del_list.name
-    del_list.delete()
-    return render(request, 'todolist/delete_successful.html', {'name': name, 'item': 'list'})
+    if user.lists.contains(del_list):
+        name = del_list.name
+        del_list.delete()
+        return render(request, 'todolist/delete_successful.html', {'name': name, 'item': 'list'})
+    else:
+        return render(request, 'todolist/access_denied.html')
 
 
 def list_create(request):
@@ -63,23 +67,30 @@ def list_create(request):
 
 
 def list_edit(request, list_id):
-    if request.method == 'POST':
-        form = ListForm(request.POST)
+    user: CustomUser = request.user
+    current_list = get_object_or_404(List, id=list_id)
+    if user.lists.contains(current_list):
+        if request.method == 'POST':
+            form = ListForm(request.POST)
 
-        if form.is_valid():
-            same_list = get_object_or_404(List, id=list_id)
-            same_list.name = form.cleaned_data['name']
-            same_list.description = form.cleaned_data['description']
-            same_list.tasks.set(form.cleaned_data['tasks'])
-            same_list.save()
-            return redirect('todolist:list_detail', list_id=list_id)
+            if form.is_valid():
+                same_list = get_object_or_404(List, id=list_id)
+                same_list.name = form.cleaned_data['name']
+                same_list.description = form.cleaned_data['description']
+                same_list.tasks.set(form.cleaned_data['tasks'])
+                same_list.save()
+                return redirect('todolist:list_detail', list_id=list_id)
 
-    elif request.method == 'GET':
-        prev_list = get_object_or_404(List, id=list_id)
-        form = ListForm({'name': prev_list.name, 'description': prev_list.description, 'tasks': prev_list.tasks.all()})
-        return render(request, 'todolist/list_form_edit.html', {"form": form, "list_id": list_id})
+        elif request.method == 'GET':
+            prev_list = get_object_or_404(List, id=list_id)
+            form = ListForm(
+                {'name': prev_list.name, 'description': prev_list.description, 'tasks': prev_list.tasks.all()})
+            return render(request, 'todolist/list_form_edit.html', {"form": form, "list_id": list_id})
 
-    return render(request, 'todolist/wrong_method.html')
+        return render(request, 'todolist/wrong_method.html')
+
+    else:
+        return render(request, 'todolist/access_denied.html')
 
 
 def task_detail(request, task_id):
